Modifications to run .NET Object allocation tool for Fibonacci number for n = 50

+++ CsharpCode/playing_with_perf_profiler/FibonacciSeries.cs
@@ -2,7 +2,7 @@
 
 internal static class FibonacciSeries
 {
-    public static int GetFiboNumber(int order)
+    public static long GetFiboNumber(int order)
     {
         if (order == 0)
         {
